refactor(eligibility): implement eligibility checker with validation requirements and integrate into frequency switch process
